Fixed issue on sending incorrect messages on last player bankrupt
Now sends up to the end of the round, including the '+ playerNo 0'
message at the end to indicate that a player has no coins left.

Used a 'sendMessages' boolean to determine whether the round has been
completed and no more messages should be added to the message queue.

Didn't use the 'gameStarted' boolean because messages must be added to
the queue BEFORE the game starts (and 'gameStarted' is false).