Removed problematic check in dealer if player is able to dd (read full)
When player inputs "doubledown" the client class already returns a
formaterror if the player is unable to (also added a check that the
number of cards in the player's hand is equal to two). There is a
redundant check in the dealer class during the player's turn for that
when currentMove == 'D'. The bug is that since there was no check in the
client class for the number of cards in the player's hand, typing
doubledown would actually change currentMove to 'D' without a
formaterror, but throw a format error in the dealer class and then loop
back through the while loop and broadcast that it is the player's turn
again. Tbh we dont even need that while loop as now it will never loop.

Also added a small delay between the dealer switching actions (ex.
dealing then checking) in case other threads dont update player's cards,
hand value etc. fast enough before the dealer moves to the next step.