Fixed a really stupid and gamebreaking bug (you should read below)
Jerry was reporting a bug that happens sometimes where the money before
and after a round stays the same despite busting.

What Jerry and Alex's A.I. groups did for some turns was spam hit or
stand. Like send "hit" then send "stand" real quick right after, for
some apparent reason I didnt bother to listen to. I realized that
actually fked up our code, because in that split second, it changes the
player's currentMove to stand. So since the player busts for hitting,
the dealer will subtract the money from their balance, but after the
dealer's turn, it also compares the player's hand value to the dealer's
hand value (since the player's currentMove variable is 'S') and
bascically compares like 22 or 23 or something to whatever the dealer
had and wins. Because of this, the player lost money from busting but
gained it back from winning that final comparison.

Although it didn't fix the problem entirely, I bypassed the problem and
functionally fixed it by simply changing the currentBet amount back to 0
immediately after any bust or blackjack, so when it also compares to the
dealer after with the stand it will just add 0 for winning. I did only
this cause I didn't want to change the code too much and potentially
break it.

Night <3